Update tests in RefreshAccessTokenTest

+++ app/src/main/kotlin/com/kondenko/pocketwaka/utils/date/DateProvider.kt
@@ -16,6 +16,6 @@ class DateProvider {
 
     fun getCurrentTimeMillis() = System.currentTimeMillis()
 
-    fun getCurrentTimeSec() = TimeUnit.MILLISECONDS.toSeconds(getCurrentTimeMillis()).toFloat()
+    fun getCurrentTimeSec() = TimeUnit.MILLISECONDS.toSeconds(getCurrentTimeMillis())
 
 }
